fixed int32 error

+++ Weather-Tracker-2/Weather-Tracker-2/Models/WeatherViewModel.cs
@@ -38,7 +38,7 @@ namespace WeatherForecast.Models
     {
         public double temp { get; set; }
         public double feels_like { get; set; }
-        public int temp_min { get; set; }
+        public double temp_min { get; set; }
         public double temp_max { get; set; }
         public int pressure { get; set; }
         public int humidity { get; set; }
